xtrs: Add support for the Lowe Electronics LE18
This seems to have been one of the more obscure graphics interfaces
possibly only sold in the UK (by Lowe Electronics) and Australia/NZ
(via Dick Smith ?)

However Lowe is a Welsh company so how could I resist.

The code is fairly minimal because apart from the addressing it's
much like any other 384x192 TRS80 overlaying video adapter. That
is 6bits wide RAM, 6bits per character across, and one entry for
each row down. In this case the two displays XOR (which means you
can also use it to just fill the graphics RAM with FF and run in
inverse video)

There is an optional EPROM with the software preloaded but that
is not emulated.

For more detail see:

http://knut.one/HI-RES_Instruction_manual.zip

Note: xtrs doesn't yet emulate the 32/64 behaviour of the Genie
which is different to the TRS80 so this code doesn't either. On
a Genie 32 column mode shows the *left* half of each video line and
there is a hardware button that lets you switch left/right.

To enable use "-le18"