xtrs: Enabling for the HyperMem plus related bug fixes to printer/maps
- Fix map 2 on the Model 4/4P: it was using the upper not lower
  bank of alternate memory. Very messy when you run an OS that
  used this map mode.
- Report the low bits of the control register in the low bits of
  printer status. HyperDrive relies upon this
- Add support for the HyperMem card on the Model 4

The Hypermem uses port 90H to select which 64K bank appears as the
upper 64K bank. Bits 4-1 provide a selection for up to 16 64K banks
(1MByte), whilst bit 0 remains used for the sound port as before.

The Hyperdrive software hard codes all support within the driver,
rather than supporting @BANK. In a way it has to because any app
making a sound is going to write 0/1 or 0/255 to the banking
register and make a fine mess of it.

There doesn't appear to be any specific way to detect the device.
Even the official ramdisc driver simply tries bank switching with
SuperMem (43H) and with Hypermem (90H) and sees if it switches the
ram at 0x8000.

I don't have a copy of HyperSVC or Megamem/Megadrive so I don't know
if there are other features only used by these and the later boards.